models/engine/file_storage.py
#!/usr/bin/python3

"""
FileStorage module
Handles serialization and deserialization of instances to and from a JSON file.
"""

import json
from typing import Dict, Any
from models.base_model import BaseModel

class FileStorage:
    """Serializes instances to a JSON file and deserializes JSON file to instances"""

    __file_path = "file.json"
    __objects: Dict[str, BaseModel] = {}

    def all(self) -> Dict[str, BaseModel]:
        """Returns the dictionary __objects"""
        return self.__objects

    def new(self, obj: BaseModel) -> None:
        """Sets in __objects the obj with key <obj class name>.id"""
        key = f"{obj.__class__.__name__}.{obj.id}"
        self.__objects[key] = obj

    def save(self) -> None:
        """Serializes __objects to the JSON file"""
        obj_dict = {key: obj.to_dict() for key, obj in self.__objects.items()}
        with open(self.__file_path, 'w') as f:
            json.dump(obj_dict, f)

    def reload(self) -> None:
        """Deserializes the JSON file to __objects"""
        try:
            with open(self.__file_path, 'r') as f:
                obj_dict = json.load(f)
                for key, value in obj_dict.items():
                    cls_name = value["__class__"]
                    if cls_name == "BaseModel":
                        self.__objects[key] = BaseModel(**value)
        except FileNotFoundError:
            pass
